# openapi.yaml
openapi: 3.0.3
info:
  title: API MongoDB — Lista de Exercícios
  version: 1.0.0
  description: |-
    Lista de Exercícios de MongoDB implementada como API REST com **Node.js + Express**
    e o driver oficial `mongodb` (sem ODM).

    Cada grupo de rotas corresponde a uma parte da lista:

    | Recurso | Exercícios | Conceitos |
    |---|---|---|
    | Produtos | 2 a 11 | CRUD, `$gt`/`$lt`, `$and`/`$or`/`$nor`, projeção, `sort`/`limit`, `$set`, `$inc` |
    | Clientes | 12, 13, 14 | documentos aninhados, `$push`, `$pull`, `$addToSet` |
    | Filmes | 15 | modelagem com array + documento aninhado |
    | Alunos | 16 | arrays de valores, `$avg`, `$unwind` |
    | Vendas | 17, 18, 19 | `$match`, `$group`, `$project`, `$multiply`, `$sum`, `$avg` |
    | Pedidos | 20 | array de documentos, `$unwind`, agrupamento por status |
    | Exercícios | todos | enunciado + comando no mongosh + resultado do banco |

    A rota **`/api/exercicios`** devolve, para cada item da lista, o enunciado, o comando
    equivalente no `mongosh` e o resultado real vindo do banco.
  contact:
    name: Repositório do trabalho
    url: http://localhost:3000
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Ambiente local
...
components:
  schemas:
    ObjectId:
      type: string
      pattern: ^[0-9a-fA-F]{24}$
      description: Identificador único do documento (24 caracteres hexadecimais).
      example: 6a98c80f3e8030cb5731e645
...
    Endereco:
      type: object
      properties:
        cidade:
          type: string
          example: Vitória da Conquista
        estado:
          type: string
          example: BA
        cep:
          type: string
          example: 45000-100
    Cliente:
      type: object
      required:
        - nome
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        nome:
          type: string
          example: Ana Silva
        email:
          type: string
          format: email
          example: ana@email.com
        endereco:
          $ref: '#/components/schemas/Endereco'
        interesses:
          type: array
          items:
            type: string
          example:
            - games
            - programacao
            - musica
        tecnologias:
          type: array
          items:
            type: string
          example:
            - JavaScript
            - MongoDB
...
    ListaFiltrada:
      type: object
      description: 'Resposta padrão das consultas: repete o filtro aplicado e devolve os dados.'
      properties:
        filtro:
          type: object
          description: Filtro montado a partir da query string.
        total:
          type: integer
          example: 5
        dados:
          type: array
          items:
            type: object
...
    ResultadoRemocao:
      type: object
      properties:
        removidos:
          type: integer
          example: 1
...
    Erro:
      type: object
      properties:
        erro:
          type: string
          example: Recurso não encontrado.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: ObjectId do documento.
      schema:
        $ref: '#/components/schemas/ObjectId'
...
  responses:
    RequisicaoInvalida:
      description: Dados inválidos ou campos obrigatórios ausentes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    NaoEncontrado:
      description: Documento não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
...
paths:
...
  /api/clientes:
    get:
      tags:
        - Clientes
      summary: Lista clientes filtrando por campo aninhado ou por array
      description: >
        `?cidade=` usa a notação de ponto (`endereco.cidade`) — exercício 12b. `?interesse=` e `?tecnologia=` buscam um
        valor dentro de um array, sem precisar de operador especial — exercícios 13a e 13b.
      parameters:
        - name: nome
          in: query
          schema:
            type: string
          example: Ana Silva
        - name: cidade
          in: query
          schema:
            type: string
          example: Vitória da Conquista
          description: Filtra por `endereco.cidade` (documento aninhado).
        - name: estado
          in: query
          schema:
            type: string
          example: BA
        - name: interesse
          in: query
          schema:
            type: string
          example: games
          description: Valor procurado dentro do array `interesses`.
        - name: tecnologia
          in: query
          schema:
            type: string
          example: MongoDB
          description: Valor procurado dentro do array `tecnologias`.
      responses:
        '200':
          description: Lista de clientes.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListaFiltrada'
                  - type: object
                    properties:
                      dados:
                        type: array
                        items:
                          $ref: '#/components/schemas/Cliente'
    post:
      tags:
        - Clientes
      summary: Cadastra um cliente (exercícios 12, 13 e 14)
      description: O corpo aceita documento aninhado (`endereco`) e arrays (`interesses`, `tecnologias`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
            examples:
              comEnderecoAninhado:
                summary: Exercício 12
                value:
                  nome: Ana Silva
                  email: ana@email.com
                  endereco:
                    cidade: Vitória da Conquista
                    estado: BA
                    cep: 45000-000
              comArray:
                summary: Exercício 13
                value:
                  nome: Bruno Santos
                  email: bruno@email.com
                  interesses:
                    - games
                    - programacao
                    - cinema
      responses:
        '201':
          description: Cliente criado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
  /api/clientes/{id}:
    get:
      tags:
        - Clientes
      summary: Busca um cliente pelo _id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Cliente encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
    delete:
      tags:
        - Clientes
      summary: Remove um cliente
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Documento removido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultadoRemocao'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /api/clientes/{id}/endereco:
    patch:
      tags:
        - Clientes
      summary: Altera campos do endereço com $set e notação de ponto (exercício 12d)
      description: >
        Cada campo enviado vira `endereco.<campo>` no `$set`, alterando apenas o subcampo. Substituir o objeto
        `endereco` inteiro apagaria os demais campos.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Endereco'
            example:
              cep: 45000-100
      responses:
        '200':
          description: Campos alterados e documento resultante.
          content:
            application/json:
              schema:
                type: object
                properties:
                  atualizou:
                    type: object
                    example:
                      endereco.cep: 45000-100
                  documento:
                    $ref: '#/components/schemas/Cliente'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /api/clientes/{id}/{campo}:
    post:
      tags:
        - Clientes
      summary: Acrescenta um valor a um array com $push ou $addToSet (exercícios 13c e 14)
      description: |
        - `unico: false` (padrão) → **$push**: acrescenta sempre, permitindo repetição.
        - `unico: true` → **$addToSet**: só acrescenta se o valor ainda não existir.

        Ao usar `$addToSet` com um valor já presente, a resposta traz o documento inalterado
        (no MongoDB, `matchedCount: 1` e `modifiedCount: 0`).
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: campo
          in: path
          required: true
          schema:
            type: string
            enum:
              - interesses
              - tecnologias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - valor
              properties:
                valor:
                  type: string
                  example: musica
                unico:
                  type: boolean
                  default: false
                  description: true usa $addToSet.
            examples:
              push:
                summary: $push (exercício 13c)
                value:
                  valor: musica
              addToSet:
                summary: $addToSet (exercício 14d)
                value:
                  valor: MongoDB
                  unico: true
      responses:
        '200':
          description: Operador aplicado e documento resultante.
          content:
            application/json:
              schema:
                type: object
                properties:
                  operador:
                    type: string
                    enum:
                      - $push
                      - $addToSet
                  valor:
                    type: string
                  documento:
                    $ref: '#/components/schemas/Cliente'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /api/clientes/{id}/{campo}/{valor}:
    delete:
      tags:
        - Clientes
      summary: Remove um valor do array com $pull (exercício 13d)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: campo
          in: path
          required: true
          schema:
            type: string
            enum:
              - interesses
              - tecnologias
        - name: valor
          in: path
          required: true
          schema:
            type: string
          example: cinema
      responses:
        '200':
          description: Documento resultante.
          content:
            application/json:
              schema:
                type: object
                properties:
                  operador:
                    type: string
                    example: $pull
                  valor:
                    type: string
                  documento:
                    $ref: '#/components/schemas/Cliente'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
...